% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{AppendClonotypeEnrichmentPVals}
\alias{AppendClonotypeEnrichmentPVals}
\title{AppendClonotypeEnrichmentPVals}
\usage{
AppendClonotypeEnrichmentPVals(dat, showProgress = FALSE)
}
\arguments{
\item{dat}{A dataframe, generated by PrepareTcrData.}

\item{showProgress}{If TRUE, a progress bar will be shown}
}
\description{
This will iterate the results from PrepareTcrData, calculating clones enriched in the activated cells relative to negative controls
}
