% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{AppendVireoData}
\alias{AppendVireoData}
\title{AppendVireoData}
\usage{
AppendVireoData(seuratObj, donorAliases = NULL, outputFieldName = "DonorId")
}
\arguments{
\item{seuratObj}{The seurat object.}

\item{donorAliases}{An optional named vector/list mapping the donorId as it appears in the vireo data (e.g., donor1) to the ID to be used in the metadata. This is currently one map for all objects, which could be problematic when merging multiple vireo datasets.}

\item{outputFieldName}{The name of the metadata field that will hold the donor assignment}
}
\description{
This will identify, download and append Vireo genetic cell assignment data for each readset in the data.
}
