% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Studies.R
\name{ApplyAcuteNxMetadata}
\alias{ApplyAcuteNxMetadata}
\title{ApplyAcuteNxMetadata}
\usage{
ApplyAcuteNxMetadata(
  seuratObj,
  errorIfUnknownIdsFound = TRUE,
  reApplyMetadata = TRUE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{errorIfUnknownIdsFound}{If true, the function will fail if the seurat object contains unknown IDs}

\item{reApplyMetadata}{If true, QueryAndApplyCdnaMetadata will be re-run}
}
\value{
A modified Seurat object.
}
\description{
Applies standard metadata related to the PPG/AcuteNx project
}
