% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{ApplyCloneFilters}
\alias{ApplyCloneFilters}
\title{ApplyCloneFilters}
\usage{
ApplyCloneFilters(
  dat,
  minCellsPerClone = 2,
  minFoldChangeAboveNoStim = 2,
  minFractionOfCloneActive = 0.01
)
}
\arguments{
\item{dat}{A dataframe, generated by PrepareTcrData}

\item{minCellsPerClone}{Any clone with fewer than this many cells in the IsActive category will be filtered}

\item{minFoldChangeAboveNoStim}{The fold change will be calculated as the fraction of cells/clone active relative to the no-stim background. Any clone below this value will be filtered}

\item{minFractionOfCloneActive}{If any clone has fewer than this fraction of cells active, it will be filtered}
}
\description{
This will filter TCR data created by PrepareTcrData
}
