% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Studies.R
\name{ApplyIMPAC_TB_Human_Metadata}
\alias{ApplyIMPAC_TB_Human_Metadata}
\title{ApplyIMPAC_TB_Human_Metadata}
\usage{
ApplyIMPAC_TB_Human_Metadata(
  seuratObj,
  reApplyMetadata = TRUE,
  errorIfUnknownIdsFound = FALSE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{reApplyMetadata}{If true, QueryAndApplyCdnaMetadata will be re-run}

\item{errorIfUnknownIdsFound}{If true, the function will fail if the seurat object contains unknown IDs}
}
\value{
A modified Seurat object.
}
\description{
Applies standard metadata related to the IMPAC_TB Study
}
