% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Studies.R
\name{ApplyPC475Metadata}
\alias{ApplyPC475Metadata}
\title{ApplyPC475Metadata}
\usage{
ApplyPC475Metadata(
  seuratObj,
  errorIfUnknownIdsFound = TRUE,
  reApplyMetadata = TRUE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{errorIfUnknownIdsFound}{If true, the function will fail if the seurat object contains unknown IDs}

\item{reApplyMetadata}{If true, QueryAndApplyCdnaMetadata will be re-run}
}
\value{
A modified Seurat object.
}
\description{
Applies standard metadata related to PC475 / BMGF HIV Reservoirs project
}
