% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{CalculateAndStoreTcrRepertioreStats}
\alias{CalculateAndStoreTcrRepertioreStats}
\title{CalculateAndStoreTcrRepertioreStats}
\usage{
CalculateAndStoreTcrRepertioreStats(seuratObj)
}
\arguments{
\item{seuratObj}{A Seurat object}
}
\value{
The dataframe with results
}
\description{
Calculates  a summary plot of clonotype data
}
