% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{CalculateClonotypeEnrichment}
\alias{CalculateClonotypeEnrichment}
\title{CalculateClonotypeEnrichment}
\usage{
CalculateClonotypeEnrichment(
  dataToTest,
  controlData,
  groupingField = "cDNA_ID",
  cloneFieldName = "Clonotype",
  contrastField = "IsActive",
  cloneSizeField = "TotalCellsForCloneAndState",
  showProgress = FALSE
)
}
\arguments{
\item{dataToTest}{A dataframe, such as generated by PrepareTcrData. This should lack the control data, with one row per clonotype}

\item{controlData}{A dataframe, such as generated by PrepareTcrData, containing the control CDR3 data, with one row per clonotype}

\item{groupingField}{A field, present in dataToTest+controlData, used to group samples.}

\item{cloneFieldName}{A field, present in dataToTest+controlData, holding the clonotype value}

\item{contrastField}{A field, present in dataToTest+controlData, used to contrast data. Typically this is isActive.}

\item{cloneSizeField}{A field, present in dataToTest+controlData, denoting the number of cells from this clonotype}

\item{showProgress}{If TRUE, a progress bar will be shown}
}
\description{
This will filter calculate enrichment of CDR3s in one population relative to a control
}
