% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAPI.R
\name{CalculateIsotype}
\alias{CalculateIsotype}
\title{CalculateIsotype}
\usage{
CalculateIsotype(seuratObj, assayName = "IG", isotypeFilterThreshold = 0.1)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{isotypeFilterThreshold}{When calculating isotype, any isotype representing below this fraction of reads in the cell is discarded. If this value is 0.1, then a cell with 5 percent of isotype reads for IHGM and 95 percent IGHA would be labeled IGHA.}
}
\value{
A modified Seurat object.
}
\description{
Uses nimble data to score isotype of each cell
}
