% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{ClassifyTNKByExpression}
\alias{ClassifyTNKByExpression}
\title{Classify T and NK By Expression}
\usage{
ClassifyTNKByExpression(seuratObj)
}
\arguments{
\item{seuratObj}{The seurat object}
}
\description{
Classify T and NK By Expression and TCR clonotype, using best available evidence of ground-truth
}
