% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodeComparison.R
\name{CompareCellBarcodeSets}
\alias{CompareCellBarcodeSets}
\title{CompareCellBarcodeSets}
\usage{
CompareCellBarcodeSets(workbooks, savePath = tempdir(), filePrefix = "")
}
\arguments{
\item{workbooks}{A vector of workbook IDs}

\item{savePath}{The folder path to use for saving output files}

\item{filePrefix}{A prefix appended to the name of all summary output files}
}
\description{
This iterates the provided workbooks, identifying every 10x cDNA library and the GEX/TCR/HTO libraries.  It generates summaries of the cell barcode intersect between them, which can help debug sample swaps.
}
