% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DownloadAndAppendCellHashing}
\alias{DownloadAndAppendCellHashing}
\title{DownloadAndAppendCellHashing}
\usage{
DownloadAndAppendCellHashing(seuratObj, outPath = tempdir())
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{outPath}{The filepath where downloaded count data will be written}
}
\value{
A modified Seurat object.
}
\description{
Downloads matching Cell Hashing data using barcodePrefix on the seurat object and appends it to metadata
}
