% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAPI.R
\name{DownloadAndAppendNimble}
\alias{DownloadAndAppendNimble}
\title{DownloadAndAppendNimble}
\usage{
DownloadAndAppendNimble(
  seuratObject,
  targetAssayName,
  outPath = tempdir(),
  enforceUniqueFeatureNames = TRUE,
  allowableGenomes = NULL,
  ensureSamplesShareAllGenomes = TRUE,
  dropAmbiguousFeatures = TRUE,
  reuseExistingDownloads = FALSE,
  performDietSeurat = TRUE,
  normalizeData = TRUE,
  assayForLibrarySize = "RNA",
  maxLibrarySizeRatio = 0.05
)
}
\arguments{
\item{seuratObject}{A Seurat object.}

\item{targetAssayName}{The target assay. If this assay exists, features will be appended (and an error thrown if there are duplicates). Otherwise a new assay will be created.}

\item{outPath}{The path to which nimble files will be downloaded and saved}

\item{enforceUniqueFeatureNames}{Whether or not to fail if we discover multiple nimble files with the same feature name}

\item{allowableGenomes}{An optional vector of genomeIds. If provided, nimble results from these genomes will be appended. If any dataset lacks a nimble file for a genome, it will fail}

\item{ensureSamplesShareAllGenomes}{If true, the function will fail unless all samples have data from the same set of genomes}

\item{dropAmbiguousFeatures}{If true, any ambiguous feature (identified as containing a comma)}

\item{reuseExistingDownloads}{If true, any pre-existing downloaded nimble TSVs will be re-used}

\item{performDietSeurat}{If true, DietSeurat will be run, which removes existing reductions. This may or may not be required based on your usage, but the default is true out of caution.}

\item{normalizeData}{Passed directly to AppendNimbleCounts()}

\item{assayForLibrarySize}{Passed directly to AppendNimbleCounts()}

\item{maxLibrarySizeRatio}{Passed directly to AppendNimbleCounts()}
}
\value{
A modified Seurat object.
}
\description{
This read a given seurat object and download/append all associated files to the object
}
