% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DownloadRawHashingDataForLoupeFile}
\alias{DownloadRawHashingDataForLoupeFile}
\title{DownloadRawHashingDataForLoupeFile}
\usage{
DownloadRawHashingDataForLoupeFile(
  loupeDataId,
  outFile,
  overwrite = T,
  countType = "raw_feature_bc_matrix"
)
}
\arguments{
\item{loupeDataId}{The outputfile Id of the loupe file}

\item{outFile}{The local path to write this file. The data will be written to a subfolder named raw_feature_bc_matrix (or filtered_feature_bc_matrix if countType=filtered_feature_bc_matrix)}

\item{overwrite}{If true, any pre-existing local copy will be replaced.}

\item{countType}{Either raw_feature_bc_matrix or filtered_feature_bc_matrix}
}
\description{
Downloads the raw_feature_bc_matrix folder for cell hashing data associated with the provided Loupe file (based on the cdna_libraries table)
}
