% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{FormatFeaturePlotColorScale}
\alias{FormatFeaturePlotColorScale}
\title{FormatFeaturePlotColorScale}
\usage{
FormatFeaturePlotColorScale(P1)
}
\arguments{
\item{P1}{The plot object}
}
\value{
A plot object
}
\description{
Calls FormatFeaturePlotColorScale and sets a navy/red color scale for Seurat FeaturePlot objects
}
