% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{FormatUmapForPresentation}
\alias{FormatUmapForPresentation}
\title{FormatUmapForPresentation}
\usage{
FormatUmapForPresentation(P1)
}
\arguments{
\item{P1}{The plot object}
}
\value{
A plot object
}
\description{
Calls renames the axes to UMAP_1 and UMAP_2 on a DimPlot/FeaturePlot
}
