% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{GenerateSRATable}
\alias{GenerateSRATable}
\title{GenerateSRATable}
\usage{
GenerateSRATable(cDNA_IDs)
}
\arguments{
\item{cDNA_IDs}{A vector of integer cDNA_IDs}
}
\description{
Returns a table with SRA-related information for a given set of cDNA_IDs
}
