% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{GenerateTcrPlot}
\alias{GenerateTcrPlot}
\title{GenerateTcrPlot}
\usage{
GenerateTcrPlot(
  dat,
  xFacetField = NA,
  plotTitle = NULL,
  yFacetField = "IsActiveLabel",
  patternField = "IsShared",
  dropInactive = FALSE,
  labelUsingCounts = TRUE,
  groupLowFreq = FALSE
)
}
\arguments{
\item{dat}{A dataframe, generated by PrepareTcrData}

\item{xFacetField}{An optional field used for faceting}

\item{plotTitle}{An optional string used as the plot title}

\item{yFacetField}{An optional field used for faceting}

\item{patternField}{An optional field used for the pattern aesthetic}

\item{dropInactive}{If true, cells where IsActive=FALSE will be dropped}

\item{labelUsingCounts}{If true, the cell total will be listed above the bars}

\item{groupLowFreq}{If true, all rows where clonotype is 'Low Freq' will be grouped to simplify plotting}
}
\description{
This will plot TCR data created by PrepareTcrData
}
