% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHC.R
\name{GroupByMhcSimilarity}
\alias{GroupByMhcSimilarity}
\title{GroupByMhcSimilarity}
\usage{
GroupByMhcSimilarity(
  seuratObj,
  groupField,
  sourceAssay = "MHC",
  dist.method = "euclidean"
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{groupField}{The field on which to group the data}

\item{sourceAssay}{The assay holding MHC data}

\item{dist.method}{Passed directly to stats::dist()}
}
\description{
Uses MHC typing data to pseudobulk data by the supplied variable and then reports a matrix of distances
}
