% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{MakeClonotypePlot}
\alias{MakeClonotypePlot}
\title{MakeClonotypePlot}
\usage{
MakeClonotypePlot(
  seuratObj,
  outFile = NULL,
  subjectId,
  chain,
  xFacetField = "Population",
  groupingFields = c("Stim", "Population", "Tissue", "SampleDate", "AssayType"),
  threshold = 0.5,
  activationFieldName = "TandNK_ActivationCore_UCell",
  lowFreqThreshold = 0.005
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{outFile}{An optional file where a TSV of the data will be written}

\item{subjectId}{The subject Id to show}

\item{chain}{The chain (i.e. TRA, TRB, TRG, TRD)}

\item{xFacetField}{Passed to facet_grid}

\item{groupingFields}{The set of fields used for grouping data}

\item{threshold}{The minimum value to consider a cell activated}

\item{activationFieldName}{The name of the field holding the score to be used to determine activation state}

\item{lowFreqThreshold}{Any clone not appearing above this threshold will be marked as low. frequency}
}
\value{
The plot object
}
\description{
Generates a summary plot of clonotype data
}
