% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAppend.R
\name{PerformMhcNormalization}
\alias{PerformMhcNormalization}
\title{PerformMhcNormalization}
\usage{
PerformMhcNormalization(
  seuratObj,
  sourceAssayName = "MHC",
  featurePrefix = "Mamu-",
  delimiter = "*",
  ambiguousFeatureDelim = ","
)
}
\arguments{
\item{sourceAssayName}{The assay to normalize}

\item{featurePrefix}{This prefix is stripped from the start of all feature names}

\item{delimiter}{Used to split the locus from allele designation}

\item{ambiguousFeatureDelim}{This character is used to split feature names in the case of ambiguous features. If a feature is ambiguous, the locus is assigned as the unique loci of the feature set.}

\item{seuratObject}{A Seurat object}
}
\value{
A modified Seurat object.
}
\description{
This is a fairly specific normalization step for MHC data. It will divide the raw counts for each feature by the sum of counts in that cell from that locus (e.g., MHC-A, MHC-B, MHC-E, MHC-I, DPA, DPB)
}
