% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{PrepareTcrData}
\alias{PrepareTcrData}
\title{PrepareTcrData}
\usage{
PrepareTcrData(
  seuratObjOrDf,
  subjectId,
  minEDS = 0,
  enforceAllDataPresent = TRUE,
  chain = "TRB",
  dropUnknownTNK_Type = FALSE,
  lowFreqThreshold = 0.001,
  retainRowsWithoutCDR3 = FALSE
)
}
\arguments{
\item{seuratObjOrDf}{Either a Seurat object or the meta.data dataframe from an object}

\item{subjectId}{The subjectId to process}

\item{minEDS}{If provided, cells with EDS less than this value will be discarded}

\item{enforceAllDataPresent}{If true, the function will error unless all stims defined in lists.tcr_stims are present in the input object}

\item{chain}{The chain to summarize}

\item{dropUnknownTNK_Type}{If true, cells with Ambiguous or Unknown TNK_Type will be dropped}

\item{lowFreqThreshold}{Clones that never have a per-sample fraction above this value will be labeled as Low Freq.}

\item{retainRowsWithoutCDR3}{If true, rather than dropping rows without CDR3 data, these will be assigned as 'No TCR' as retained}
}
\description{
This will group TCR data by the supplied chain, and calculate fields used for filtering
}
