% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell.R
\name{QueryAndApplyCdnaMetadata}
\alias{QueryAndApplyCdnaMetadata}
\title{QueryAndApplyCdnaMetadata}
\usage{
QueryAndApplyCdnaMetadata(
  seuratObj,
  fieldSelect = c("rowid", "sortid/population", "sortid/sampleid/subjectid",
    "sortid/sampleid/sampledate", "sortid/sampleid/stim", "sortid/sampleid/tissue",
    "plateid", "workbook/workbookid", "sortid/sampleid/assaytype", "status", "chemistry"),
  fieldNames = c("cDNA_ID", "Population", "SubjectId", "SampleDate", "Stim", "Tissue",
    "PlateId", "WorkbookId", "AssayType", "cDNA_Status", "LibraryChemistry"),
  overwriteExisting = TRUE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{fieldSelect}{The set of fields to query}

\item{fieldNames}{The labels to use for the fields}

\item{overwriteExisting}{If true and the output exists, it will be overwritten}
}
\value{
A modified Seurat object.
}
\description{
This will read the barcodePrefix from the seurat object, query the LK server and apply the appropriate metadata from the cDNA table
}
