% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{RenameUmapAxes}
\alias{RenameUmapAxes}
\title{RenameUmapAxes}
\usage{
RenameUmapAxes(prefix = "UMAP")
}
\arguments{
\item{prefix}{The prefix to apply to the axis labels}
}
\value{
A plot object
}
\description{
Renames the axes to UMAP_1 and UMAP_2 on a DimPlot/FeaturePlot. Usage: Dimplot() + RenameUmapAxes()
}
\examples{
\dontrun{
Seurat::DimPlot(seuratObj) +
     RenameUmapAxes()

Seurat::DimPlot(seuratObj) +
     RenameUmapAxes(prefix = 'tSNE')
}
}
