% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabKeySettings.R
\name{SetLabKeyDefaults}
\alias{SetLabKeyDefaults}
\title{SetLabKeyDefaults}
\usage{
SetLabKeyDefaults(baseUrl = "", defaultFolder = "")
}
\arguments{
\item{baseUrl}{The baseUrl of the LabKey server.}

\item{defaultFolder}{The default folder path (which should generally be the parent folder and not a workbook) for queries.}
}
\description{
Can be used to set the default LabKey baseUrl and folderPath for all queries of this module.
}
