% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{SeuratToCoNGA}
\alias{SeuratToCoNGA}
\title{SeuratToCoNGA}
\usage{
SeuratToCoNGA(seuratObj, output_dir)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{output_dir}{The local path to write the output files.}
}
\description{
A wrapper function to prepare a Seurat object for Conga.
}
