% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{SummarizeTNK_Activation}
\alias{SummarizeTNK_Activation}
\title{MakeClonotypePlot}
\usage{
SummarizeTNK_Activation(
  seuratObj,
  outFile,
  xFacetField = "Population",
  groupingFields = c("SampleDate", "Stim", "Population", "AssayType"),
  activationFieldName = "TandNK_Activation_UCell",
  threshold = 0.5
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{outFile}{The output file path to which results will be written}

\item{groupingFields}{The set of fields used for grouping data}

\item{activationFieldName}{The name of the field holding the score to be used to determine activation state}

\item{threshold}{The minimum value to consider a cell activated}
}
\value{
A modified Seurat object.
}
\description{
Generates a summary plot of clonotype data
}
