% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{UploadOutputFile}
\alias{UploadOutputFile}
\title{UploadOutputFile}
\usage{
UploadOutputFile(
  localPath,
  workbook,
  name,
  category,
  description,
  readsetId = NULL,
  genomeId,
  remoteFilename = NULL
)
}
\arguments{
\item{localPath}{The local filepath to the file}

\item{workbook}{The numeric ID of the target workbook}

\item{name}{The name for this output}

\item{category}{The category for this output}

\item{description}{A description for this output}

\item{readsetId}{The integer readsetId associated with this output}

\item{genomeId}{The integer genome ID associated with this output}

\item{remoteFilename}{If null, the local name will be used}
}
\description{
Uploads a local file and saves as an output file tracked in LabKey.
}
