% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAppend.R
\name{AppendNimbleCounts}
\alias{AppendNimbleCounts}
\title{AppendNimbleCounts}
\usage{
AppendNimbleCounts(
  seuratObject,
  nimbleFile,
  targetAssayName,
  dropAmbiguousFeatures = TRUE
)
}
\arguments{
\item{seuratObject}{A Seurat object.}

\item{nimbleFile}{A nimble file, which is a TSV of feature counts created by nimble}

\item{targetAssayName}{The target assay. If this assay exists, features will be appended (and an error thrown if there are duplicates). Otherwise a new assay will be created.}

\item{dropAmbiguousFeatures}{If true, any ambiguous features (defined as containing a comma) will be discarded}
}
\value{
A modified Seurat object.
}
\description{
Reads a given seurat object and a nimble file, and appends the nimble data to the object.
}
