% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAppend.R
\name{AppendNimbleCounts}
\alias{AppendNimbleCounts}
\title{AppendNimbleCounts}
\usage{
AppendNimbleCounts(
  seuratObject,
  nimbleFile,
  appendToExistingAssay = FALSE,
  dropAmbiguousFeatures = TRUE,
  targetAssayName = "RNA",
  newAssayName = "Nimble"
)
}
\arguments{
\item{seuratObject}{A Seurat object.}

\item{nimbleFile}{A nimble file, which is a TSV of feature counts created by nimble}

\item{appendToExistingAssay}{If true, append the nimble data to the count matrix. If false, make a new Nimble assay}

\item{dropAmbiguousFeatures}{If true, any ambiguous features (defined as containing a comma) will be discarded}

\item{targetAssayName}{If appendToExistingAssay=TRUE, nimble data will be appended to this assay}

\item{newAssayName}{If appendToExistingAssay=FALSE, nimble data will be stored in a new assay with this name}
}
\value{
A modified Seurat object.
}
\description{
Reads a given seurat object and a nimble file, and appends the nimble data to the object.
}
