% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAppend.R
\name{AppendNimbleCounts}
\alias{AppendNimbleCounts}
\title{AppendNimbleCounts}
\usage{
AppendNimbleCounts(
  seuratObj,
  nimbleFile,
  targetAssayName,
  maxAmbiguityAllowed = 0,
  renameConflictingFeatures = TRUE,
  duplicateFeatureSuffix = ".Nimble",
  normalizeData = TRUE,
  performDietSeurat = (targetAssayName \%in\% names(seuratObj@assays)),
  assayForLibrarySize = "RNA",
  maxLibrarySizeRatio = 0.05,
  doPlot = TRUE,
  maxFeaturesToPlot = 40,
  replaceExistingAssayData = TRUE,
  featureRenameList = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{nimbleFile}{A nimble file, which is a TSV of feature counts created by nimble}

\item{targetAssayName}{The target assay. If this assay exists, features will be appended (and an error thrown if there are duplicates). Otherwise a new assay will be created.}

\item{maxAmbiguityAllowed}{If provided, any features representing more than ths value will be discarded. For example, 'Feat1,Feat2,Feat3' represents 3 features. maxAmbiguityAllowed=1 results in removal of all ambiguous features.}

\item{renameConflictingFeatures}{If true, when appending to an existing assay, any conflicting feature names will be renamed, appending the value of duplicateFeatureSuffix}

\item{duplicateFeatureSuffix}{If renameConflictingFeatures is true, this string will be appended to duplicated feature names}

\item{normalizeData}{If true, data will be normalized after appending/creating the assay. This will default to CellMembrane::LogNormalizeUsingAlternateAssay; however, if assayForLibrarySize equals targetAssayName then Seurat::NormalizeData is used.}

\item{performDietSeurat}{If true, DietSeurat will be run, which removes existing reductions. This may or may not be required based on your usage, but the default is to perform this if the targetAssay exists.}

\item{assayForLibrarySize}{If normalizeData is true, then this is the assay used for librarySize when normalizing. If assayForLibrarySize equals targetAssayName, Seurat::NormalizeData is used.}

\item{maxLibrarySizeRatio}{If normalizeData is true, then this is passed to CellMembrane::LogNormalizeUsingAlternateAssay}

\item{doPlot}{If true, FeaturePlots will be generated for the appended features}

\item{maxFeaturesToPlot}{If doPlot is true, this is the maximum number of features to plot}

\item{replaceExistingAssayData}{If true, any existing data in the targetAssay will be deleted}

\item{featureRenameList}{An optional named list in the format <OLD_NAME> = <NEW_NAME>. If any <OLD_NAME> are present, the will be renamed to <NEW_NAME>. The intention of this is to recover specific ambiguous classes.}
}
\value{
A modified Seurat object.
}
\description{
Reads a given seurat object and a nimble file, and appends the nimble data to the object.
}
