% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAppend.R
\name{AppendNimbleCounts}
\alias{AppendNimbleCounts}
\title{AppendNimbleCounts}
\usage{
AppendNimbleCounts(
  seuratObject,
  nimbleFile,
  targetAssayName,
  dropAmbiguousFeatures = TRUE,
  renameConflictingFeatures = TRUE,
  duplicateFeatureSuffix = ".Nimble",
  normalizeData = TRUE,
  performDietSeurat = TRUE,
  assayForLibrarySize = "RNA",
  maxLibrarySizeRatio = 0.05
)
}
\arguments{
\item{seuratObject}{A Seurat object.}

\item{nimbleFile}{A nimble file, which is a TSV of feature counts created by nimble}

\item{targetAssayName}{The target assay. If this assay exists, features will be appended (and an error thrown if there are duplicates). Otherwise a new assay will be created.}

\item{dropAmbiguousFeatures}{If true, any ambiguous features (defined as containing a comma) will be discarded}

\item{renameConflictingFeatures}{If true, when appending to an existing assay, any conflicting feature names will be renamed, appending the value of duplicateFeatureSuffix}

\item{duplicateFeatureSuffix}{If renameConflictingFeatures is true, this string will be appended to duplicated feature names}

\item{normalizeData}{If true, data will be normalized after appending/creating the assay. This will default to CellMembrane::LogNormalizeUsingAlternateAssay; however, if assayForLibrarySize equals targetAssayName then Seurat::NormalizeData is used.}

\item{performDietSeurat}{If true, DietSeurat will be run, which removes existing reductions. This may or may not be required based on your usage, but the default is true out of caution.}

\item{assayForLibrarySize}{If normalizeData is true, then this is the assay used for librarySize when normalizing. If assayForLibrarySize equals targetAssayName, Seurat::NormalizeData is used.}

\item{maxLibrarySizeRatio}{If normalizeData is true, then this is passed to CellMembrane::LogNormalizeUsingAlternateAssay}
}
\value{
A modified Seurat object.
}
\description{
Reads a given seurat object and a nimble file, and appends the nimble data to the object.
}
