% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{ApplyKnownClonotypicData}
\alias{ApplyKnownClonotypicData}
\title{ApplyKnownClonotypicData}
\usage{
ApplyKnownClonotypicData(
  seuratObj,
  antigenInclusionList = NULL,
  antigenExclusionList = NULL,
  minActivationFrequency = 0,
  minFractionCloneActivated = 0,
  fieldPrefix = NULL,
  doNotPruneUsingCognateChain = FALSE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{antigenInclusionList}{If provided, only antigens from this list will be considered}

\item{antigenExclusionList}{If provided, antigens on this list will be omitted}

\item{minActivationFrequency}{If provided, only responses with activationFrequency (of the parent population) will be included}

\item{minFractionCloneActivated}{If provided, only responses where fractionCloneActivated is above this value will be will be included}

\item{fieldPrefix}{If provided, this will be appended to the beginning of the output field names}

\item{doNotPruneUsingCognateChain}{By default, matching is performed using the primary chain (as defined in the DB). If this is provided, this additional step is skipped}
}
\description{
This will query the clone_responses table and append a column tagging each cell for matching antigens (based on clonotype)
}
