% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{ApplyKnownClonotypicData}
\alias{ApplyKnownClonotypicData}
\title{ApplyKnownClonotypicData}
\usage{
ApplyKnownClonotypicData(seuratObj)
}
\arguments{
\item{seuratObj}{The seurat object}
}
\description{
This will query the clone_responses table and append a column tagging each cell for matching antigens (based on clonotype)
}
