% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{CalculateAndStoreTcrRepertoireStats}
\alias{CalculateAndStoreTcrRepertoireStats}
\title{CalculateAndStoreTcrRepertoireStats}
\usage{
CalculateAndStoreTcrRepertoireStats(seuratObj, outputFile = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{outputFile}{An optional filepath, where the complete results will be written as a TSV}
}
\value{
The dataframe with results
}
\description{
Calculates  a summary plot of clonotype data
}
