% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{ConvertAxesToArrows}
\alias{ConvertAxesToArrows}
\title{ConvertAxesToArrows}
\usage{
ConvertAxesToArrows(
  arrowLength = 0.3,
  arrowLinewidth = 1.5,
  arrowHeadLenMm = 2,
  arrowHeadAngle = 50,
  arrowLabelOffsetX = 0.08,
  arrowLabelOffsetY = 0.02,
  arrowYlabelAngle = 90,
  renameUmapAxes = TRUE,
  plotAxisPrefix = "UMAP"
)
}
\arguments{
\item{arrowLength}{Numeric. Shaft length as a fraction of the axis range.}

\item{arrowLinewidth}{Numeric. Line width of the arrow shafts.}

\item{arrowHeadLenMm}{Numeric. Arrowhead length in millimeters.}

\item{arrowHeadAngle}{Numeric. Arrowhead angle in degrees (larger = “bigger” head).}

\item{arrowLabelOffsetX}{Numeric. Offset of labels from the arrow line as a fraction of the x axis range (vertical for x label, horizontal for y label).}

\item{arrowLabelOffsetY}{Numeric. Offset of labels from the arrow line as a fraction of the x axis range (vertical for x label, horizontal for y label).}

\item{arrowYlabelAngle}{Numeric. Rotation angle (degrees) for the y-axis arrow label.}

\item{renameUmapAxes}{If true, RenameUmapAxes() will be run}

\item{plotAxisPrefix}{Passed directly to RenameUmapAxes(prefix = plotAxisPrefix)}

\item{xLabel}{Character. Text label placed at the midpoint of the x-axis arrow.}

\item{yLabel}{Character. Text label placed at the midpoint of the y-axis arrow.}
}
\value{
ggplot object
}
\description{
Replaces standard x/y axis lines with short, arrow-based axes and
  adds labels at their midpoints (useful for UMAPs and other DR plots).
}
\examples{
\dontrun{
Seurat::DimPlot(seuratObj) +
     ConvertAxesToArrows()

Seurat::DimPlot(seuratObj) +
     ConvertAxesToArrows(plotAxisPrefix = 'tSNE')
}
}
