% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{ConvertAxesToArrows}
\alias{ConvertAxesToArrows}
\title{ConvertAxesToArrows}
\usage{
ConvertAxesToArrows(
  arrowLength = 0.3,
  arrowLinewidth = 1.5,
  arrowHeadLenMm = 2,
  arrowHeadAngle = 50,
  renameUmapAxes = TRUE,
  plotAxisPrefix = "UMAP"
)
}
\arguments{
\item{arrowLength}{Numeric. Shaft length as a fraction of the axis range.}

\item{arrowLinewidth}{Numeric. Line width of the arrow shafts.}

\item{arrowHeadLenMm}{Numeric. Arrowhead length in millimeters.}

\item{arrowHeadAngle}{Numeric. Arrowhead angle in degrees (larger = “bigger” head).}

\item{renameUmapAxes}{If true, RenameUmapAxes() will be run}

\item{plotAxisPrefix}{Passed directly to RenameUmapAxes(prefix = plotAxisPrefix)}
}
\value{
ggplot object
}
\description{
Replaces standard x/y axis lines with short, arrow-based axes and
  adds labels at their midpoints (useful for UMAPs and other DR plots).
}
\examples{
\dontrun{
Seurat::DimPlot(seuratObj) +
     ConvertAxesToArrows()

Seurat::DimPlot(seuratObj) +
     ConvertAxesToArrows(plotAxisPrefix = 'tSNE')
}
}
