% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAPI.R
\name{DownloadAndAppendNimble}
\alias{DownloadAndAppendNimble}
\title{DownloadAndAppendNimble}
\usage{
DownloadAndAppendNimble(
  seuratObject,
  outPath = tempdir(),
  enforceUniqueFeatureNames = TRUE,
  allowableGenomes = NULL
)
}
\arguments{
\item{outPath}{The path to which nimble files will be downloaded and saved}

\item{enforceUniqueFeatureNames}{Whether or not to fail if we discover multiple nimble files with the same feature name}

\item{seuratObj}{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
This read a given seurat object and download/append all associated files to the object
}
