% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DownloadMetadataForSeuratObject}
\alias{DownloadMetadataForSeuratObject}
\title{DownloadMetadataForSeuratObject}
\usage{
DownloadMetadataForSeuratObject(
  outputFileId,
  outFile,
  overwrite = TRUE,
  returnDataFrame = FALSE,
  deleteFile = returnDataFrame,
  showProgressBar = FALSE
)
}
\arguments{
\item{outputFileId}{The rowid of the sequence outputfiles on the webserver}

\item{outFile}{The local path to write this file}

\item{overwrite}{If true, any pre-existing local copy will be replaced.}

\item{returnDataFrame}{If true, the metadata will be read to a data.frame and returned}

\item{deleteFile}{If true, the outFile will be deleted. This only makes sense when used with returnDataFrame=TRUE}

\item{showProgressBar}{If true, a progress bar will be shown}
}
\description{
Downloads an output file tracked in LabKey to the local filesystem.
}
