% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{FormatFeaturePlotColorScale}
\alias{FormatFeaturePlotColorScale}
\title{FormatFeaturePlotColorScale}
\usage{
FormatFeaturePlotColorScale(prefix = "UMAP")
}
\arguments{
\item{prefix}{The string prefix for the axes. Passed to RenameUmapAxes()}
}
\value{
A plot object
}
\description{
Calls FormatFeaturePlotColorScale and sets a navy/red color scale for Seurat FeaturePlot objects
}
\examples{
\dontrun{
Seurat::DimPlot(seuratObj) +
     FormatFeaturePlotColorScale()
}
}
