% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{FormatFeaturePlotColorScale}
\alias{FormatFeaturePlotColorScale}
\title{FormatFeaturePlotColorScale}
\usage{
FormatFeaturePlotColorScale()
}
\value{
A plot object
}
\description{
Calls FormatFeaturePlotColorScale and sets a navy/red color scale for Seurat FeaturePlot objects
}
\examples{
\dontrun{
Seurat::FeaturePlot(seuratObj, features = 'CD8A') &
     FormatFeaturePlotColorScale()
}
}
