% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{GenerateTcrPlot}
\alias{GenerateTcrPlot}
\title{GenerateTcrPlot}
\usage{
GenerateTcrPlot(
  dat,
  xFacetField = NA,
  yFacetField = "IsActiveLabel",
  dropInactive = FALSE
)
}
\arguments{
\item{dat}{A dataframe, generated by PrepareTcrData}

\item{xFacetField}{An optional field used for faceting}

\item{yFacetField}{An optional field used for faceting}

\item{subjectId}{The subjectId to process}
}
\description{
This will plot TCR data created by PrepareTcrData
}
