% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{GroupOverlappingClones}
\alias{GroupOverlappingClones}
\title{GroupOverlappingClones}
\usage{
GroupOverlappingClones(dat, groupingFields, dataMask = NULL)
}
\arguments{
\item{dat}{A dataframe, generated by PrepareTcrData}

\item{groupingFields}{A list of fields, typically sample-variables, on which to group}

\item{dataMask}{An optional logical vector. If provided, the 'dat' dataframe is filtered to include only rows where dataMask is TRUE.}
}
\description{
This will plot TCR data created by PrepareTcrData
}
