% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{GroupOverlappingClones}
\alias{GroupOverlappingClones}
\title{GroupOverlappingClones}
\usage{
GroupOverlappingClones(
  dat,
  groupingFields,
  maxRatioToCombine = 0.5,
  dataMask = NULL
)
}
\arguments{
\item{dat}{A dataframe, generated by PrepareTcrData}

\item{groupingFields}{A list of fields, typically sample-variables, on which to group}

\item{maxRatioToCombine}{If provided, the function will calculate the total number of cells in the parent group divided by the total number of cells in the group to combine. It will only merge groups is that ratio is above this value}

\item{dataMask}{An optional logical vector. If provided, the 'dat' dataframe is filtered to include only rows where dataMask is TRUE.}
}
\description{
This will plot TCR data created by PrepareTcrData
}
