% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcrStims.R
\name{IdentifyAndStoreActiveClonotypes}
\alias{IdentifyAndStoreActiveClonotypes}
\title{IdentifyAndStoreActiveClonotypes}
\usage{
IdentifyAndStoreActiveClonotypes(
  seuratObj,
  chain = "TRB",
  method = "sPLS",
  storeStimLevelData = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object. This expects}

\item{chain}{The chain to summarize}

\item{method}{Either 'Cluster-Based' or 'sPLS'}

\item{storeStimLevelData}{If true, activation levels will be stord in tcrtb.stims}
}
\description{
This is a very specialized method designed to score and summarize activated TCR clonotypes
}
