% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAPI.R
\name{PerformDefaultNimbleAppend}
\alias{PerformDefaultNimbleAppend}
\title{PerformDefaultNimbleAppend}
\usage{
PerformDefaultNimbleAppend(
  seuratObj,
  maxLibrarySizeRatio = 100,
  assayForLibrarySize = "RNA"
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{maxLibrarySizeRatio}{Passed directly to AppendNimbleCounts()}

\item{assayForLibrarySize}{Passed directly to AppendNimbleCounts()}
}
\value{
A modified Seurat object.
}
\description{
This is designed to wrap a series of nimble download commands into one, along with some domain-specific logic for each type of data
}
