% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleAppend.R
\name{PerformMhcNormalization}
\alias{PerformMhcNormalization}
\title{PerformMhcNormalization}
\usage{
PerformMhcNormalization(
  seuratObj,
  sourceAssayName = "MHC",
  featurePrefix = "Mamu-",
  delimiter = "*",
  ambiguousFeatureDelim = ",",
  perCell = TRUE,
  cellGroupingVariable = "DatasetId",
  stripNumbersFromLocus = TRUE
)
}
\arguments{
\item{sourceAssayName}{The assay to normalize}

\item{featurePrefix}{This prefix is stripped from the start of all feature names}

\item{delimiter}{Used to split the locus from allele designation}

\item{ambiguousFeatureDelim}{This character is used to split feature names in the case of ambiguous features. If a feature is ambiguous, the locus is assigned as the unique loci of the feature set.}

\item{perCell}{If true, the feature counts are scaled based on the library size of features from that locus in that cell. If false, it is scaled based on the library size of features in that locus from all cells matching cellGroupingVariable}

\item{cellGroupingVariable}{If perCell is FALSE, the library size is calculated by taking the sum of features from that locus across all cells where this metadata variable matches the current cell}

\item{stripNumbersFromLocus}{If true, numeric values will be stripped from all locus strings}

\item{seuratObject}{A Seurat object}
}
\value{
A modified Seurat object.
}
\description{
This is a fairly specific normalization step for MHC data. It will divide the raw counts for each feature by the sum of counts in that cell from that locus (e.g., MHC-A, MHC-B, MHC-E, MHC-I, DPA, DPB)
}
